refactor: 사다리 관련 클래스 패키지화
