+++ src/main/java/ladder/domain/LadderGame.java
@@ -1,6 +1,8 @@
 package ladder.domain;
 
 import ladder.domain.creator.LadderCreator;
+import ladder.domain.ladder.Ladder;
+import ladder.domain.ladder.LadderHeight;
 
 public class LadderGame {
     private final LadderCreator ladderCreator;
