+++ src/main/java/ladder/domain/creator/LadderCreator.java
@@ -1,6 +1,6 @@
 package ladder.domain.creator;
 
-import ladder.domain.Ladder;
+import ladder.domain.ladder.Ladder;
 
 public interface LadderCreator {
     Ladder create(int width, int height);
